requirements.txt
pyTelegramBotAPI
requests
urllib3
pytz
google-generativeai
gradio_client
pydantic
pydantic-settings
colorlog
